vix_project/requirements.txt
vaderSentiment
pandas
plotly
streamlit
requests
python-dotenv
pyTelegramBotAPI
yfinance
